Move cloud debug executable into clouddebug package
