+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/clouddebug/CloudDebugExecutable.kt
@@ -1,10 +1,14 @@
-// Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package software.aws.toolkits.jetbrains.core.executables
+package software.aws.toolkits.jetbrains.services.clouddebug
 
 import com.intellij.openapi.util.SystemInfo
 import com.intellij.util.text.SemVer
+import software.aws.toolkits.jetbrains.core.executables.AutoResolvable
+import software.aws.toolkits.jetbrains.core.executables.ExecutableCommon
+import software.aws.toolkits.jetbrains.core.executables.ExecutableType
+import software.aws.toolkits.jetbrains.core.executables.Validatable
 
 import software.aws.toolkits.jetbrains.settings.ExecutableDetector
 import java.nio.file.Path
