+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/clouddebug/CloudDebugVersionCache.kt
@@ -1,11 +1,12 @@
-// Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package software.aws.toolkits.jetbrains.core.executables
+package software.aws.toolkits.jetbrains.services.clouddebug
 
 import com.intellij.execution.process.CapturingProcessHandler
 import com.intellij.util.text.SemVer
 import com.intellij.util.text.nullize
+import software.aws.toolkits.jetbrains.core.executables.ExecutableCommon
 import software.aws.toolkits.jetbrains.utils.FileInfoCache
 import software.aws.toolkits.resources.message
